fix(postmark): Do not create the json manually
Previously, json was created manually which resulted in characters
like quotes not getting escaped properly. This resulted in invalid
json sent to the server causing errors. Now a seperate
DataContractJsonSerializer is used to serialize user input into the
appropriate json for the server.

Some additional fixes/updates too:

* When options is "long", rules come back from the api which are an
array of rules. Previously the code assumed it was a single string.
This has been updated, to be an array of SpamcheckRule so it gets
deserialized correctly.
* Adding nuget specific attributes to the csproj so this can be
submitted as a nuget package.